+++ core/src/main/resources/META-INF/services/io.questdb.griffin.FunctionFactory
@@ -891,7 +891,6 @@ io.questdb.griffin.engine.functions.groupby.VarGroupByFunctionFactory
 
 # var_samp()
 io.questdb.griffin.engine.functions.groupby.VarSampleGroupByFunctionFactory
-io.questdb.griffin.engine.functions.groupby.ApproxVarSampleGroupByFunctionFactory
 
 # var_pop()
 io.questdb.griffin.engine.functions.groupby.VarPopGroupByFunctionFactory
